MapBottomSheet Component
